refactor: 네이밍 규칙에 맞게 명칭 수정
