+++ src/main/java/com/nighttrip/core/domain/user/entity/User.java
@@ -2,8 +2,8 @@ package com.nighttrip.core.domain.user.entity;
 
 import com.nighttrip.core.domain.avatar.entity.Avatar;
 import com.nighttrip.core.domain.tripplan.entity.TripPlan;
-import com.nighttrip.core.global.enums.Oauth_Provider;
-import com.nighttrip.core.global.enums.User_role;
+import com.nighttrip.core.global.enums.OauthProvider;
+import com.nighttrip.core.global.enums.UserRole;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -32,7 +32,7 @@ public class User {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
-    private Oauth_Provider provider;
+    private OauthProvider provider;
 
     @Column(name = "social_id", nullable = false)
     private String socialId;
@@ -42,7 +42,7 @@ public class User {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
-    private User_role role = User_role.USER;
+    private UserRole role = UserRole.USER;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "avatar_id")
@@ -51,7 +51,7 @@ public class User {
     @OneToMany(mappedBy = "user")
     private List<TripPlan> tripPlans = new ArrayList<>();
 
-    public User(String email, String nickname, String socialId, Oauth_Provider provider) {
+    public User(String email, String nickname, String socialId, OauthProvider provider) {
         this.email = email;
         this.nickname = nickname;
         this.socialId = socialId;
